[TASK-200] userData 수정
