+++ src/components/MainHeader/MainHeader.tsx
@@ -1,27 +1,10 @@
 import { HeaderLogo } from '@images/index';
 import styles from './MainHeader.module.scss';
-import { useEffect } from 'react';
-import { tentenInstance } from '@constants/axios';
-import userDataAtom from '@recoil/userData';
-import { useRecoilState } from 'recoil';
 import { TRAVEL_MINI_ICON, TravelType } from '@constants/testResult';
+import useGetUserData from '@src/apis/useGetUserData';
 
 const MainHeader = ({ handleOpen }: { handleOpen: () => void }) => {
-  const [userData, setUserData] = useRecoilState(userDataAtom);
-
-  useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const response = await tentenInstance.get('/users/current');
-
-        setUserData(response.data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchUserData();
-  }, []);
+  const { userData } = useGetUserData();
 
   return (
     <div className={styles.wrapper}>
